Make Create patch a view component
*This component is now used both in create and edit mode.
